linera_chain
Module data_types

Structs

BlockExecutionOutcome

The messages and the state hash resulting from a ProposedBlock’s execution.

BlockProposal

An authenticated proposal for a new block.

ChainAndHeight

A chain ID with a block height.

IncomingBundle

A bundle of cross-chain messages.

LiteValue

The hash and chain ID of a CertificateValue.

LiteVote

A vote on a statement from a validator, represented as a LiteValue.

MessageBundle

A set of messages from a single block, for a single destination.

OperationResult

The execution result of a single operation.

PostedMessage

A message together with kind, authentication and grant information.

ProposalContent

The data a block proposer signs.

ProposedBlock

A block containing operations to apply on a given chain, as well as the acknowledgment of a number of incoming messages from other chains.

SignatureAggregator

Vote

A vote on a statement from a validator.

Enums

MessageAction

What to do with a message picked from the inbox.

OriginalProposal

An earlier proposal that is being retried.

Transaction

A transaction in a block: incoming messages or an operation.

Traits

OutgoingMessageExt