Struct linera_chain::ChainStateView

pub struct ChainStateView<C> {
    pub execution_state: ExecutionStateView<C>,
    pub execution_state_hash: RegisterView<C, Option<CryptoHash>>,
    pub tip_state: RegisterView<C, ChainTipState>,
    pub manager: RegisterView<C, ChainManager>,
    pub confirmed_log: LogView<C, CryptoHash>,
    pub received_log: LogView<C, ChainAndHeight>,
    pub communication_states: CollectionView<C, ApplicationId, CommunicationStateView<C>>,
}

A view accessing the state of a chain.

Fields
execution_state: ExecutionStateView<C>

Execution state, including system and user applications.

execution_state_hash: RegisterView<C, Option<CryptoHash>>

Hash of the execution state.

tip_state: RegisterView<C, ChainTipState>

Block-chaining state.

manager: RegisterView<C, ChainManager>

Consensus state.

confirmed_log: LogView<C, CryptoHash>

Hashes of all certified blocks for this sender. This ends with block_hash and has length usize::from(next_block_height).

received_log: LogView<C, ChainAndHeight>

Sender chain and height of all certified blocks known as a receiver (local ordering).

communication_states: CollectionView<C, ApplicationId, CommunicationStateView<C>>

Communication state of applications.

Implementations

impl<C> ChainStateView<C>where
C: Context + Clone + Send + Sync + 'static,
ViewError: From<C::Error>,
C::Extra: ExecutionRuntimeContext,

pub fn chain_id(&self) -> ChainId

pub async fn query_application(
&mut self,
application_id: ApplicationId,
query: &Query
) -> Result<Response, ChainError>

pub async fn describe_application(
&mut self,
application_id: UserApplicationId
) -> Result<UserApplicationDescription, ChainError>

pub async fn mark_messages_as_received(
&mut self,
application_id: ApplicationId,
target: Target,
height: BlockHeight
) -> Result<bool, ChainError>

pub fn is_active(&self) -> bool

Invariant for the states of active chains.

pub fn ensure_is_active(&self) -> Result<(), ChainError>

Invariant for the states of active chains.

pub async fn validate_incoming_messages(&mut self) -> Result<(), ChainError>

Verify that this chain is up-to-date and all the messages executed ahead of time have been properly received by now.

pub async fn next_block_height_to_receive(
&mut self,
application_id: ApplicationId,
origin: Origin
) -> Result<BlockHeight, ChainError>

pub async fn last_anticipated_block_height(
&mut self,
application_id: ApplicationId,
origin: Origin
) -> Result<Option<BlockHeight>, ChainError>

pub async fn receive_block(
&mut self,
application_id: ApplicationId,
origin: &Origin,
height: BlockHeight,
timestamp: Timestamp,
effects: Vec<OutgoingEffect>,
certificate_hash: CryptoHash
) -> Result<(), ChainError>

Schedule operations to be executed as a recipient, unless this block was already processed. Returns true if the call changed the chain state. Operations must be received by order of heights and indices.

pub async fn remove_events_from_inboxes(
&mut self,
block: &Block
) -> Result<(), ChainError>

Removes the incoming messages in the block from the inboxes.

pub async fn execute_block(
&mut self,
block: &Block
) -> Result<Vec<OutgoingEffect>, ChainError>

Execute a new block: first the incoming messages, then the main operation.

Modifies the state of inboxes, outboxes, and channels, if needed.

As usual, in case of errors, self may not be consistent any more and should be thrown away.

Returns the list of effects caused by the block being executed.