linera_chain

Module types

Convenience re-exports of the public block and certificate types.

Re-exports

pub use super::block::*;

Structs

ConfirmedBlockCertificate

Certificate for a ConfirmedBlock instance, certified in some round by a quorum of ConfirmedBlock votes (which carry no unlocking round).

GenericCertificate

Generic type representing a certificate for value of type T.

LiteCertificate

A certified statement from the committee, without the value.

ValidatedBlockCertificate

Certificate for a ValidatedBlock instance, certified in some round whose ValidatedBlock voters signed an unlocking round.

Enums

Certificate

Enum wrapping all types of certificates that can be created. A certified statement from the committee. Every certificate is a statement signed by the quorum of the committee.

CertificateKind

The kind of value a certificate certifies.

Traits

CertificateValue

A value that can be certified by a quorum of validators.

Certified

The common read interface shared by all certificate types: the signed value, the round and unlocking round it was certified under, its signatures, and verification.

Type Aliases

TimeoutCertificate

Certificate for a Timeout instance. A timeout certificate means that the next consensus round has begun.