Struct linera_chain::types::GenericCertificate

pub struct GenericCertificate<T> {
    pub round: Round,
    /* private fields */
}

Generic type representing a certificate for value of type T.

Fields
round: Round
Implementations

impl GenericCertificate<ConfirmedBlock>

pub fn executed_block(&self) -> &ExecutedBlock

Returns reference to the ExecutedBlock contained in this certificate.

pub fn has_message(&self, message_id: &MessageId) -> bool

Returns whether this value contains the message with the specified ID.

pub fn message_bundles_for<'a>( &'a self, medium: &'a Medium, recipient: ChainId, ) -> impl Iterator<Item = (Epoch, MessageBundle)> + 'a

Returns the bundles of messages sent via the given medium to the specified recipient. Messages originating from different transactions of the original block are kept in separate bundles. If the medium is a channel, does not verify that the recipient is actually subscribed to that channel.

pub fn requires_blob(&self, blob_id: &BlobId) -> bool

pub fn outgoing_message_count(&self) -> usize

impl<T> GenericCertificate<T>

pub fn new( value: Hashed<T>, round: Round, signatures: Vec<(ValidatorName, Signature)>, ) -> Self

pub fn value(&self) -> &Hashed<T>

Returns a reference to the Hashed value contained in this certificate.

pub fn into_value(self) -> Hashed<T>

Consumes this certificate, returning the value it contains.

pub fn inner(&self) -> &T

Returns reference to the value contained in this certificate.

pub fn into_inner(self) -> T

Consumes this certificate, returning the value it contains.

pub fn hash(&self) -> CryptoHash

Returns the certified value’s hash.

pub fn destructure(self) -> (Hashed<T>, Round, Vec<(ValidatorName, Signature)>)

pub fn signatures(&self) -> &Vec<(ValidatorName, Signature)>

pub fn signatures_mut(&mut self) -> &mut Vec<(ValidatorName, Signature)>

pub fn add_signature( &mut self, signature: (ValidatorName, Signature), ) -> &Vec<(ValidatorName, Signature)>

Adds a signature to the certificate’s list of signatures It’s the responsibility of the caller to not insert duplicates

pub fn is_signed_by(&self, validator_name: &ValidatorName) -> bool

Returns whether the validator is among the signatories of this certificate.

pub fn check(&self, committee: &Committee) -> Result<(), ChainError>
where T: CertificateValue,

Verifies the certificate.

pub fn lite_certificate(&self) -> LiteCertificate<'_>
where T: CertificateValue,

impl GenericCertificate<ValidatedBlock>

pub fn requires_blob(&self, blob_id: &BlobId) -> bool

pub fn outgoing_message_count(&self) -> usize

pub fn executed_block(&self) -> &ExecutedBlock

Returns reference to the ExecutedBlock contained in this certificate.