Type Alias linera_chain::types::Certificate

pub type Certificate = GenericCertificate<CertificateValue>;

A certified statement from the committee.

Aliased Type

struct Certificate {
    pub round: Round,
    /* private fields */
}

Fields
round: Round
Implementations

impl Certificate

pub fn lite_certificate(&self) -> LiteCertificate<'_>

Returns the certificate without the full value.

pub fn lite_value(&self) -> LiteValue

Returns the LiteValue corresponding to the certified value.

pub fn message_bundles_for<'a>( &'a self, medium: &'a Medium, recipient: ChainId, ) -> impl Iterator<Item = (Epoch, MessageBundle)> + 'a

Returns the bundles of messages sent via the given medium to the specified recipient. Messages originating from different transactions of the original block are kept in separate bundles. If the medium is a channel, does not verify that the recipient is actually subscribed to that channel.

pub fn requires_blob(&self, blob_id: &BlobId) -> bool

pub fn outgoing_message_count(&self) -> usize