linera_chain

Module block

Structs

Block

Block defines the atomic unit of growth of the Linera chain.

BlockBody

The body of a block containing all the data included in the block.

BlockHeader

Succinct representation of a block. Contains all the metadata to follow the chain of blocks or verifying inclusion (event, message, oracle response, etc.) in the block’s body.

ConfirmedBlock

Wrapper around a Block that has been confirmed.

PreviousEventBlocksMap

PreviousMessageBlocksMap

Timeout

ValidatedBlock

Wrapper around a Block that has been validated.

Enums

ConversionError

Failure to convert a Certificate into one of the expected certificate types.