Module linera_chain::data_types

Structs

Block

A block containing operations to apply on a given chain, as well as the acknowledgment of a number of incoming messages from other chains.

BlockExecutionOutcome

The messages and the state hash resulting from a Block’s execution.

BlockProposal

An authenticated proposal for a new block.

ChainAndHeight

A chain ID with a block height.

ChannelFullName

A channel name together with its application ID.

EventRecord

An event recorded in an executed block.

ExecutedBlock

A Block, together with the outcome from its execution.

IncomingBundle

A bundle of cross-chain messages.

LiteValue

The hash and chain ID of a CertificateValue.

LiteVote

A vote on a statement from a validator, represented as a LiteValue.

MessageBundle

A set of messages from a single block, for a single destination.

Origin

The origin of a message, relative to a particular application. Used to identify each inbox.

OutgoingMessage

A posted message together with routing information.

PostedMessage

A message together with kind, authentication and grant information.

ProposalContent

The data a block proposer signs.

SignatureAggregator

Target

The target of a message, relative to a particular application. Used to identify each outbox.

Vote

A vote on a statement from a validator.

Enums

Medium

The origin of a message coming from a particular chain. Used to identify each inbox.

MessageAction

What to do with a message picked from the inbox.

Transaction

A transaction in a block: incoming messages or an operation.