Module linera_chain::data_types

Structs

Block

A block containing operations to apply on a given chain, as well as the acknowledgment of a number of incoming messages from other chains.

BlockAndRound

A block with a round number.

BlockProposal

An authenticated proposal for a new block.

Certificate

A certified statement from the committee.

CertificateWithDependencies

A certificate, together with others required for its execution.

ChainAndHeight

A chain ID with a block height.

ChannelFullName

A channel name together with its application id.

Event

An effect together with non replayable information to ensure uniqueness in a particular inbox.

HashedValue

A statement to be certified by the validators, with its hash.

LiteCertificate

A certified statement from the committee, without the value.

LiteValue

The hash and chain ID of a Value.

LiteVote

A vote on a statement from a validator, represented as a LiteValue.

Message

A message received from a block of another chain.

Origin

The origin of a message, relative to a particular application. Used to identify each inbox.

OutgoingEffect

An effect together with routing information.

SignatureAggregator

Target

The target of a message, relative to a particular application. Used to identify each outbox.

Value

A statement to be certified by the validators.

Vote

A vote on a statement from a validator.

Enums

Medium

The origin of a message coming from a particular chain. Used to identify each inbox.

ValueKind

The type of a value: whether it’s a validated block in a particular round, or already confirmed.