linera_chain

Module types

Re-exports

pub use super::block::*;

Structs

GenericCertificate

Generic type representing a certificate for value of type T.

LiteCertificate

A certified statement from the committee, without the value.

Enums

Certificate

Enum wrapping all types of certificates that can be created. A certified statement from the committee. Every certificate is a statement signed by the quorum of the committee.

CertificateKind

Traits

CertificateValue

Type Aliases

ConfirmedBlockCertificate

Certificate for a ConfirmedBlock instance. A confirmed block certificate means that the block is finalized: It is the agreed block at that height on that chain.

TimeoutCertificate

Certificate for a Timeout instance. A timeout certificate means that the next consensus round has begun.

ValidatedBlockCertificate

Certificate for a ValidatedBlock instance. A validated block certificate means the block is valid (but not necessarily finalized yet). Since only one block per round is validated, there can be at most one such certificate in every round.