linera_chain::types

Type Alias ValidatedBlockCertificate

pub type ValidatedBlockCertificate = GenericCertificate<ValidatedBlock>;

Certificate for a ValidatedBlock instance. A validated block certificate means the block is valid (but not necessarily finalized yet). Since only one block per round is validated, there can be at most one such certificate in every round.

Aliased Type

struct ValidatedBlockCertificate {
    pub round: Round,
    /* private fields */
}

Fields
round: Round